going for 1000 votes (#11)
* going for 1000 votes

* a few tweaks

---------

+++ off-chain/src/main/java/org/cardanofoundation/hydrapoc/model/ChallengeProposal.java
@@ -1,3 +1,3 @@
 package org.cardanofoundation.hydrapoc.model;
 
-public record ChallengeProposal(long challenge, long proposal) {}
+public record ChallengeProposal(int challenge, int proposal) {}
